util: Refactor device service logic
